# ZKVM-1333: Improve Executor Segment Creation Performance