Before, every time the executor creates a Segment, it cloned the whole MemoryImage which is expensive and then it rehashes everything as well.  

Before:
```
┌─────────┬────────┬────────────┬──────────┬─────────┬─────┬──────┐
│ name    │ hashfn │ throughput │ duration │ cycles  │ ram │ seal │
├─────────┼────────┼────────────┼──────────┼─────────┼─────┼──────┤
│ execute │ N/A    │ 40.8MHz    │ 24.3ms   │ 1014.1K │ N/A │ N/A  │
└─────────┴────────┴────────────┴──────────┴─────────┴─────┴──────┘
```
After:
```
┌─────────┬────────┬────────────┬──────────┬─────────┬─────┬──────┐
│ name    │ hashfn │ throughput │ duration │ cycles  │ ram │ seal │
├─────────┼────────┼────────────┼──────────┼─────────┼─────┼──────┤
│ execute │ N/A    │ 44.5MHz    │ 22.3ms   │ 1014.1K │ N/A │ N/A  │
└─────────┴────────┴────────────┴──────────┴─────────┴─────┴──────┘
```

I've ran this against zeth as well ` RUST_LOG="info" RISC0_INFO=1 target/release/zeth-ethereum run --cache=cache --block-number=22176200 --block-count=1 `

Before:
```
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::exec::executor] execution time: 27.017512174s
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::session] number of segments: 575
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::session] 602931200 total cycles
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::session] 502496192 user cycles (83.34%)
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::session] 96775344 paging cycles (16.05%)
[2025-04-16T21:17:07Z INFO  risc0_zkvm::host::server::session] 3659664 reserved cycles (0.61%)
```

After:
```
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::exec::executor] execution time: 23.589958846s
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::session] number of segments: 575
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::session] 602931200 total cycles
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::session] 502555065 user cycles (83.35%)
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::session] 96868690 paging cycles (16.07%)
[2025-04-16T22:00:01Z INFO  risc0_zkvm::host::server::session] 3507445 reserved cycles (0.58%)
```