@ec2 2025-04-21
After making the page hashing done in parallel via rayon, I've managed to get `zeth` down to ~18s. 

```
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::exec::executor] execution time: 18.582866867s
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::session] number of segments: 588
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::session] 616562688 total cycles
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::session] 502406956 user cycles (81.49%)
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::session] 96793698 paging cycles (15.70%)
[2025-04-21T21:19:54Z INFO  risc0_zkvm::host::server::session] 17362034 reserved cycles (2.82%)```