@ec2 2025-04-21
@bobbobbio Calculating partial images on a separate thread indeed was a speed up. However, we were cloning the whole `MemoryImage` and then sending it to another thread for calculation. As time goes on, each `MemoryImage` becomes larger and larger and thus more expensive... 

In addition, the way the partial image was being constructed in the separate thread was not very efficient because what we were doing was rehashing EVERYTHING to create this partial image even though the `MemoryImage` we are using already has the hashes calculated at that point.